OPSMN-3020: Should display number of objects in the list page. - Added os-list-pager widget to display total count of records.
- Behaviour of os-list-pager widget:
  > If total records > 100, it will show "Showing 100 of many more records".
  > Clicking on many more link will get actual count of total records, once many more link is clicked,
    it will show total count for all applied filter for current page.
  > If total records <= 100 say 55, it will show "Showing 55 records".

- Added APIs to disaply total count for below list of entities:
  1.Collection Protocol
  2.Participant
  3.Institutes
  4.Users
  5.Sites
  6.Containers
  7.Container Types
  8.Specimen Carts
  9.Forms
  10.Distribution Protocols
  11.Distribution Orders
  12.Shipments
  13.Jobs

+++ WEB-INF/src/com/krishagni/catissueplus/core/administrative/repository/StorageContainerDao.java
@@ -19,7 +19,7 @@ public interface StorageContainerDao extends Dao<StorageContainer> {
 
 	public Map<String, Object> getContainerIds(String key, Object value);
 
-	public int getStorageContainersCount(StorageContainerListCriteria listCrit);
+	public Long getStorageContainersCount(StorageContainerListCriteria listCrit);
 
 	public List<String> getNonCompliantContainers(ContainerRestrictionsCriteria crit);
